+++ index.html
@@ -235,8 +235,7 @@
       <h2 class="data-error__title">Не удалось загрузить данные</h2>
     </section>
   </template>
-  <script scr="js/functions.js"></script>
-  <script src="js/main.js" defer></script>
+  <script src="js/main.js" type="module"></script>
 </body>
 
 </html>
